BLD/RF: Big restructuring. All modules back inside the main pelita module.
