+++ pelita/players/FoodEatingPlayer.py
