+++ pelita/players/SmartEatingPlayer.py
