+++ pelita/players/SmartRandomPlayer.py
