+++ pelita/scripts/pelita_player.py
@@ -106,7 +106,7 @@ def load_factory(filespec):
 
 def import_builtin_player(name):
     with pelita.utils.with_sys_path("./"):
-        players_module = __import__("players")
+        players_module = __import__('pelita.players', fromlist='players')
         sane_players = {p.__name__: p for p in players_module.SANE_PLAYERS}
 
     if name == 'random':
