+++ pelita/tournament/komode.py
