--- tournament/test.yaml
@@ -1,21 +1,0 @@
----
-location: Pelipolis
-date: 2525
-team_prefix: group
-seed: null
-bonusmatch: True
-teams:
-  - spec: RandomPlayer
-    members:
-      - "Member One"
-  - spec: StoppingPlayer
-    members:
-      - "I stop"
-  - spec: SmartRandomPlayer
-    members:
-      - "I am smart"
-  - spec: FoodEatingPlayer
-    members:
-      - "no-one"
-    id: group3
-
